Prevent NullPointerException after clicking "Show more" button
